Move to .Internal namespaces
