+++ src/EFCore.Cosmos.Sql/Extensions/CosmosSqlMetadataExtensions.cs
@@ -1,9 +1,10 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using Microsoft.EntityFrameworkCore.Cosmos.Sql.Metadata;
 using Microsoft.EntityFrameworkCore.Metadata;
 
-namespace Microsoft.EntityFrameworkCore.Cosmos.Sql.Metadata
+namespace Microsoft.EntityFrameworkCore
 {
     public static class CosmosSqlMetadataExtensions
     {
