+++ src/EFCore.Cosmos.Sql/Metadata/Internal/CosmosSqlAnnotationNames.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.EntityFrameworkCore.Cosmos.Sql.Metadata
+namespace Microsoft.EntityFrameworkCore.Cosmos.Sql.Metadata.Internal
 {
     public static class CosmosSqlAnnotationNames
     {
