+++ src/EFCore.Cosmos.Sql/Query/Internal/CosmosSqlQueryContextFactory.cs
@@ -2,12 +2,10 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Collections.Generic;
-using System.Text;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Query;
 
-namespace Microsoft.EntityFrameworkCore.Cosmos.Sql.Query
+namespace Microsoft.EntityFrameworkCore.Cosmos.Sql.Query.Internal
 {
     public class CosmosSqlQueryContextFactory : QueryContextFactory
     {
