+++ test/EFCore.Cosmos.Sql.FunctionalTests/EFCore.Cosmos.Sql.FunctionalTests.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFrameworks>$(StandardTestTfms)</TargetFrameworks>
     <AssemblyName>Microsoft.EntityFrameworkCore.Cosmos.Sql.FunctionalTests</AssemblyName>
-    <RootNamespace>Microsoft.EntityFrameworkCore</RootNamespace>
+    <RootNamespace>Microsoft.EntityFrameworkCore.Cosmos.Sql</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
