+++ test/EFCore.Cosmos.Sql.Tests/EFCore.Cosmos.Sql.Tests.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFrameworks>$(StandardTestTfms)</TargetFrameworks>
     <AssemblyName>Microsoft.EntityFrameworkCore.Cosmos.Sql.Tests</AssemblyName>
-    <RootNamespace>Microsoft.EntityFrameworkCore</RootNamespace>
+    <RootNamespace>Microsoft.EntityFrameworkCore.Cosmos.Sql</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
