+++ test/EFCore.Cosmos.Sql.Tests/TestUtilities/CosmosSqlTestHelpers.cs
@@ -2,9 +2,10 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using Microsoft.EntityFrameworkCore.TestUtilities;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Microsoft.EntityFrameworkCore.TestUtilities
+namespace Microsoft.EntityFrameworkCore.Cosmos.Sql.TestUtilities
 {
     public class CosmosSqlTestHelpers : TestHelpers
     {
